requirements/ai_requirements.txt
# AI Integration Requirements for YOUTILITY5
# NOTE: This file has optional AI dependencies. Base requirements are in base.txt
# Some packages may have compatibility issues on macOS - see documentation

# txtai - All-in-one AI framework for semantic search and LLM orchestration
txtai[all]>=7.0.0
txtai[embeddings]>=7.0.0
txtai[pipeline]>=7.0.0

# MindsDB - AI database automation platform
# Updated to latest version for scikit-learn 1.5.x compatibility
# TEMPORARILY DISABLED: Conflicts with Pillow>=11.3.0 security requirement (CVE-2025-48379)
# Reason: mindsdb pulls in langchain-nvidia-ai-endpoints which requires pillow<11.0.0
# Since NVIDIA endpoints are not used in the codebase, prioritizing security over optional AI feature
# To install MindsDB manually if needed: pip install mindsdb>=25.9.0 (downgrades Pillow to 10.x)
# mindsdb>=25.9.0  # Previous versions (23.12.4.0) had scikit-learn conflicts
# mindsdb_datasources>=1.0.0

# Vector databases and similarity search
# faiss-cpu>=1.7.4  # Removed - already in base.txt
# faiss-gpu>=1.7.4  # Uncomment for GPU support
# chromadb>=0.4.15  # Removed - using pgvector instead
hnswlib>=0.7.0    # Hierarchical Navigable Small World graphs

# Machine Learning and Data Science
# NOTE: These are already in base.txt - duplicates removed to avoid conflicts
# scikit-learn>=1.5.0  # Already in base.txt line 207
# numpy>=1.24.0  # Already in base.txt line 139
# pandas>=2.0.0  # Already in base.txt line 159
# scipy>=1.11.0  # Already in base.txt line 208
xgboost>=2.0.0  # Gradient boosting for fraud detection and imbalanced classification
shap>=0.45.0    # SHAP explainability for ML model interpretability (Phase 4)

# Deep Learning and Transformers
# NOTE: These are already in base.txt - duplicates removed to avoid conflicts
# torch>=2.7.2  # Already in base.txt line 234 (pinned to 2.8.0)
# transformers>=4.53.0  # Already in base.txt line 237 (pinned to 4.53.0)
sentence-transformers>=2.2.2
accelerate>=0.24.0

# Voice Recognition and Audio Processing
# CRITICAL: Required for voice biometric authentication
resemblyzer>=0.1.1.dev0  # Speaker recognition and voice verification
librosa>=0.10.0          # Audio analysis and feature extraction
pydub>=0.25.1            # Audio manipulation and format conversion
soundfile>=0.12.1        # Audio I/O (read/write audio files)
webrtcvad>=2.0.10        # Voice activity detection
silero-vad>=4.0.0        # Modern VAD using neural networks (optional enhancement)

# Natural Language Processing
# nltk>=3.9  # Already in base.txt line 138 (pinned to 3.9.1)
# spacy>=3.8.6  # Already in base.txt line 204 (updated to 3.8.6+ for Python 3.13)
textblob>=0.17.1

# Time Series Forecasting (for MindsDB integration)
statsforecast>=1.4.0
neuralforecast>=1.6.0
hierarchicalforecast>=0.3.0

# OpenAI and LLM integrations
openai>=1.45.0
anthropic>=0.7.0  # For Claude integration (backup/fallback)
google-generativeai>=0.8.3  # Google Gemini (primary LLM for agent intelligence)
# langchain>=0.1.0  # REMOVED: Not used in codebase (Parlant is used instead)
# langchain-community>=0.0.10  # REMOVED: Not used in codebase
# Parlant SDK is optional and currently pins nonexistent OpenTelemetry builds.
# Install manually if needed once upstream pins are fixed.
# parlant>=3.0.0

# Database and Caching
redis>=5.0.0      # For caching and real-time features
celery>=5.3.0     # For background task processing
# psycopg2-binary removed - use psycopg[binary,pool]==3.2.7 from base.txt instead
pgvector>=0.2.4   # PostgreSQL vector extension

# API and Web Framework enhancements
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-extensions>=3.2.3
django-debug-toolbar>=4.2.0

# Monitoring and Logging
structlog>=23.1.0
django-prometheus>=2.3.1

# Data Processing and ETL
dask>=2023.9.0    # For parallel processing
joblib>=1.3.0     # For model persistence
pyarrow>=13.0.0   # For efficient data serialization

# Visualization and Charts (for dashboard widgets)
plotly>=5.17.0
matplotlib>=3.7.0
seaborn>=0.12.0

# HTTP and API clients
requests>=2.32.4  # Security update: CVE-2024-47081
httpx==0.28.1     # Modern HTTP client (compatible with parlant/openai)
aiohttp>=3.9.0    # Async HTTP client

# Configuration and Environment
python-decouple>=3.8
pydantic>=2.5.0   # For data validation
python-dotenv>=1.0.0

# Testing and Development
pytest>=7.4.0
pytest-django>=4.7.0
pytest-asyncio>=0.21.0
factory-boy>=3.3.0
faker>=19.12.0

# Security and Authentication
cryptography>=44.0.1  # Security update: CVE-2024-12797
PyJWT>=2.8.0

# File Processing
python-magic>=0.4.27
Pillow>=11.3.0  # Security update: CVE-2025-48379
pypdf>=3.17.0     # For PDF processing

# Async and Concurrency
asyncio-mqtt>=0.16.1  # For real-time notifications
websockets>=12.0      # For real-time dashboard updates

# Development and Debugging
ipython>=8.15.0
ipdb>=0.13.13
black>=23.9.0     # Code formatting
isort>=5.12.0     # Import sorting
flake8>=6.1.0     # Code linting

# Documentation
sphinx>=7.1.0
sphinx-rtd-theme>=1.3.0

# Additional utilities
python-dateutil>=2.8.2
pytz>=2023.3
tqdm>=4.66.0      # Progress bars
rich>=13.6.0      # Rich console output
